+++ docs/pages/_en/1.0/reference/release-notes.md
@@ -14,6 +14,7 @@ group: Deployment & Configs
 
 - Added 'View Crop Region' component.
 - Added 'Add Section View' component.
+- Added 'Add Detail View' component.
 
 
 {% endcapture %}